Reorganize framework-like code into a framework directory. (#1216)
* Reorganize framework-like code into a framework directory.

* Actually move API tests
