+++ index.yaml
@@ -2,165 +2,198 @@ indexes:
 
 # AUTOGENERATED
 
-# This index.yaml is automatically updated whenever the Cloud Datastore
-# emulator detects that a new type of query is run. If you want to manage the
-# index.yaml file manually, remove the "# AUTOGENERATED" marker line above.
-# If you want to manage some indexes manually, move them above the marker line.
+# This index.yaml is automatically updated whenever the dev_appserver
+# detects that a new type of query is run.  If you want to manage the
+# index.yaml file manually, remove the above marker line (the line
+# saying "# AUTOGENERATED").  If you want to manage some indexes
+# manually, move them above the marker line.  The index.yaml file is
+# automatically uploaded to the admin console when you next deploy
+# your application using appcfg.py.
 
-- kind: "AnimatedProperty"
+- kind: AnimatedProperty
   properties:
-  - name: "bucket_id"
-  - name: "date"
-- kind: "AnimatedProperty"
+  - name: bucket_id
+  - name: date
+
+- kind: AnimatedProperty
   properties:
-  - name: "bucket_id"
-  - name: "date"
+  - name: bucket_id
+  - name: date
     direction: desc
-- kind: "AnimatedProperty"
+
+- kind: AnimatedProperty
   properties:
-  - name: "date"
-  - name: "day_percentage"
+  - name: date
+  - name: day_percentage
     direction: desc
-- kind: "Feature"
+
+- kind: Feature
   properties:
-  - name: "category"
-  - name: "updated"
+  - name: category
+  - name: updated
     direction: desc
-- kind: "Feature"
+
+- kind: Feature
   properties:
-  - name: "category"
-  - name: "updated"
+  - name: category
+  - name: updated
     direction: desc
-  - name: "name"
-- kind: "Feature"
+  - name: name
+
+- kind: Feature
   properties:
-  - name: "deleted"
-  - name: "updated"
+  - name: deleted
+  - name: updated
     direction: desc
-- kind: "Feature"
+
+- kind: Feature
   properties:
-  - name: "impl_status_chrome"
-  - name: "name"
-- kind: "Feature"
+  - name: impl_status_chrome
+  - name: name
+
+- kind: Feature
   properties:
-  - name: "impl_status_chrome"
-  - name: "name"
+  - name: impl_status_chrome
+  - name: name
     direction: desc
-- kind: "Feature"
+
+- kind: Feature
   properties:
-  - name: "impl_status_chrome"
-  - name: "shipped_milestone"
-  - name: "name"
-- kind: "Feature"
+  - name: impl_status_chrome
+  - name: shipped_milestone
+  - name: name
+
+- kind: Feature
   properties:
-  - name: "impl_status_chrome"
-  - name: "shipped_milestone"
+  - name: impl_status_chrome
+  - name: shipped_milestone
     direction: desc
-  - name: "name"
-- kind: "Feature"
+  - name: name
+
+- kind: Feature
   properties:
-  - name: "impl_status_chrome"
-  - name: "shipped_milestone"
+  - name: impl_status_chrome
+  - name: shipped_milestone
     direction: desc
-  - name: "name"
+  - name: name
     direction: desc
-- kind: "Feature"
+
+- kind: Feature
   properties:
-  - name: "impl_status_chrome"
+  - name: impl_status_chrome
     direction: desc
-  - name: "shipped_milestone"
+  - name: shipped_milestone
     direction: desc
-  - name: "name"
-- kind: "Feature"
+  - name: name
+
+- kind: Feature
   properties:
-  - name: "name"
-  - name: "updated"
-- kind: "Feature"
+  - name: name
+  - name: updated
+
+- kind: Feature
   properties:
-  - name: "name"
-  - name: "updated"
+  - name: name
+  - name: updated
     direction: desc
-- kind: "Feature"
+
+- kind: Feature
   properties:
-  - name: "sample_links"
-  - name: "updated"
+  - name: sample_links
+  - name: updated
     direction: desc
-- kind: "Feature"
+
+- kind: Feature
   properties:
-  - name: "shipped_milestone"
-  - name: "name"
-- kind: "Feature"
+  - name: shipped_milestone
+  - name: name
+
+- kind: Feature
   properties:
-  - name: "shipped_milestone"
-  - name: "shipped_android_milestone"
+  - name: shipped_milestone
+  - name: shipped_android_milestone
     direction: desc
-  - name: "name"
-- kind: "Feature"
+  - name: name
+
+- kind: Feature
   properties:
-  - name: "shipped_milestone"
+  - name: shipped_milestone
     direction: desc
-  - name: "name"
-- kind: "Feature"
+  - name: name
+
+- kind: Feature
   properties:
-  - name: "shipped_milestone"
+  - name: shipped_milestone
     direction: desc
-  - name: "name"
-  - name: "impl_status_chrome"
-- kind: "Feature"
+  - name: name
+  - name: impl_status_chrome
+
+- kind: Feature
   properties:
-  - name: "shipped_milestone"
+  - name: shipped_milestone
     direction: desc
-  - name: "name"
-  - name: "impl_status_chrome"
+  - name: name
+  - name: impl_status_chrome
     direction: desc
-- kind: "Feature"
+
+- kind: Feature
   properties:
-  - name: "updated"
+  - name: updated
     direction: desc
-  - name: "name"
-- kind: "FeatureObserver"
+  - name: name
+
+- kind: FeatureObserver
   properties:
-  - name: "bucket_id"
-  - name: "date"
-- kind: "FeatureObserver"
+  - name: bucket_id
+  - name: date
+
+- kind: FeatureObserver
   properties:
-  - name: "bucket_id"
-  - name: "date"
+  - name: bucket_id
+  - name: date
     direction: desc
-- kind: "FeatureObserver"
+
+- kind: FeatureObserver
   properties:
-  - name: "date"
-  - name: "day_percentage"
+  - name: date
+  - name: day_percentage
     direction: desc
-- kind: "FeatureOwner"
+
+- kind: FeatureOwner
   properties:
-  - name: "blink_components"
-  - name: "name"
-- kind: "FeatureOwner"
+  - name: blink_components
+  - name: name
+
+- kind: FeatureOwner
   properties:
-  - name: "primary_blink_components"
-  - name: "name"
-- kind: "StableInstance"
+  - name: primary_blink_components
+  - name: name
+
+- kind: StableInstance
   properties:
-  - name: "bucket_id"
-  - name: "date"
-- kind: "StableInstance"
+  - name: bucket_id
+  - name: date
+
+- kind: StableInstance
   properties:
-  - name: "bucket_id"
-  - name: "date"
+  - name: bucket_id
+  - name: date
     direction: desc
-- kind: "StableInstance"
+
+- kind: StableInstance
   properties:
-  - name: "bucket_id="
-  - name: "date"
+  - name: bucket_id=
+  - name: date
     direction: desc
-- kind: "StableInstance"
+
+- kind: StableInstance
   properties:
-  - name: "date"
-  - name: "day_percentage"
+  - name: date
+  - name: day_percentage
     direction: desc
-- kind: "StableInstance"
+
+- kind: StableInstance
   properties:
-  - name: "date"
-  - name: "hits"
+  - name: date
+  - name: hits
     direction: desc
