+++ models.py
@@ -26,12 +26,11 @@ import time
 from google.appengine.ext import db
 # from google.appengine.ext.db import djangoforms
 from google.appengine.api import mail
-import ramcache
+from framework import ramcache
 import requests
 from google.appengine.api import users
 
-import cloud_tasks_helpers
-import common
+from framework import cloud_tasks_helpers
 import settings
 import util
